refactor: jsx 파일에서 직접 경로 대신 import 방식으로 아이콘 불러오도록 수정
