+++ src/pages/Intro/Intro.jsx
@@ -5,7 +5,7 @@ const Intro = memo(() => {
   return (
     <div className="intro-container">
       <div className="netflix-logo">
-        <img src="./src/assets/netflix-logo-icon.svg" alt="Netflix Logo" />
+        <img src="./src/assets/logo.svg" alt="Netflix Logo" />
       </div>
     </div>
   );
