+++ src/pages/Search/Search.css
@@ -29,6 +29,8 @@
   height: 20px;
   background: url(../../assets/icon-search.svg) no-repeat 0 0;
   border: 0;
+  filter: invert(60%) sepia(0%) saturate(0%) hue-rotate(0deg) brightness(100%)
+    contrast(91%);
 }
 .suggestions {
   position: absolute;
